Added ManifestPlugin and shared Ordering
The JPMSModuleInfoPlugin used the VerifierPlugin and then changed the set of resources. This is not a good idea since those resources will then not be in the manifest name section.

This primarily adds a Manifest Plugin which is called at manifest creation. It is called after the main section is done but before the name section. There is also a method that is called after the main section.

Since both needed to be ordered, the ordered part is extracted.

---
 Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>

Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>