Resolve changes
The resolve now allows all types to be seen by the resolver. This used to be filtered to only bundle or fragment. However, to support the idea of SupportingResource, they must remain in the resources. The RunResolution is  properly adapted to filter the non-bundle/fragment resources.

This should be invisible to most but everybody should be aware that people that use the API directly might run in incompatibilities.

---
 Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>

Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>