Backward compatibility for the resource filtering To make the MRJ scenario work, we needed synthetic resource. However, for some reason I do not get, we filtered them out before the resolver saw them. The filtering was moved to the final stage when we filter the bundles. However, this means that there are theoretical situations where people need the old behavior. Release 7 is a major release so we can break but it is clearly not nice to have a backup
Therefore this change adds a -resolve.reject instruction. This is a requirement list that
is applied to the capability. If it matches, the capability is removed. To be able to mimic
the 6.x behavior, a requirement can be prefixed with a , which indicates that
a capability should be skipped when its resource has a capability that matches the requirement.

Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>