Implement MR Bundle indexing using multiple bundle resources
Rather than using virtual resources this commit makes use of real osgi.bundle resources to represent a Multi Release Bundle. This has the advantage that no additional filtering is required when resolving. One significant difficulty is that both Felix and bnd rely heavily on the Resource equals method, and any two reasources with matching URI/SHA information in their content capabilities are considered equal. This has the potential to break any scenario where multiple resources are backed by the same physical bytes. This commit therefore adds a fragment to the osgi.content capability urls for eachadditional resource, preventing them from being seen as equal. This fragment is required not to be sent to a server and is transparently removed when using URL openStream.

Signed-off-by: Tim Ward <timothyjward@apache.org>