gradle: Use an ArtifactView for jar library elements selection
Using an ArtifactView avoids ordering issues when the Jar task is
realized prior to the compileClasspath configuration being modified
just before resolution.  It makes the attributes used to resolve
libraries more predictable, and prevents an internal behavior change in
Gradle 8.2 from affecting this process.

This also avoids modifying attributes on Gradle-created configurations,
which should be avoided.

Signed-off-by: Tom Tresansky <ttresansky@gradle.com>
Signed-off-by: BJ Hargrave <bj@hargrave.dev>