Maven release archive option
The archive option used the context given in the RepositoryPlugin.put
as the base processor. This is true for the release process but
the `-buildrepo` option had a special install method in Project. This
method created a processor per file. This copied processor did
not have the same file base nor did it collect the warnings and
errors.

---
 Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>

Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>