Add indirection to fetch the Workspace instead of using static singleton
Currently there is a lot of code that is bound to the "Central" class
what has several problems when it comes to reuse this outside of
bndtools.

This adds two ways of getting out of this reference chain for components
that should be reusable:

1) From a project context use the Eclipse Adapter Pattern to adapt an
IProject > Workspace
2) For cases where a global workspace model is desired and no project
context exits, the "central" Workspace is provided as a service in the
runntime.

Signed-off-by: Christoph Läubrich <laeubi@laeubi-soft.de>