pass test for GAV
1st draft, base for discussion.
I hooked into the indexing of BridgeRepository because commons-cli:commons-cli:1.2 (commons-cli-1.2.jar) is indexed as commons-cli:commons-cli:1.2.0 by bridge repo (notice 1.2 vs. 1.2.0). To satisfy the existing testcase of MavenBndRepo this was the only place where I could get the version as 3 digit semver. To be discussed if it really should be that way

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>