Add a verify goal to the bnd-resolver-maven-plugin
The  goal provides a way to validate that a set of run bundles is a valid resolution. It does this by creating a set of run requirements requiring each run bundle by identity and performing a resolve operation. This resolve operation forbids any capabilities that come from bundles not in the list of run requirements. If the resolution succeeds then it is further checked against the orginal run requirements.

If these checks pass then we know that:

 * The resolution is valid
 * The resolution satisfies the original run requirements

This is useful for CI when we want to test that a checked-in bndrun file has been correctly resolved.

Fixes #5984

Signed-off-by: Tim Ward <timothyjward@apache.org>