add Repo action Update revisions to higher MICRO
this basically does what the MbrCommand does, but in the UI. You can update a central.mvn index file with this action.
There is a DryRun option, which just copies the result to clipboard without touching the index file. you can use that to simulate it and check the result upfront.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>