set resolution loglevel based on -resolvedebug
previously resolving used the ResolverLogger() default constructor. This commit checks if -resolvedebug instruction has been set e.g. in a .bndrun file and if yes it uses the value from it and initializes the ResolveLogger with it.
also we now print the resolvedebug log to the resolutionFailurePanel. This also shows the logfile location on disk if the log is too large for display.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>