avoid resolving default to absolute path
the ${.} in the default-value caused to expand the value always to an absolute path on the harddrive of the developer editing and saving the file. if you accidentally commit this, then it wil fail on other machines.
Let's just define the name, relative to the build.bnd . it can still be changed manually

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>