do not expand here-macro ${.} in the .bnd UI editor
With this change ensures that macros are not expanded when they are edited in the UI e.g. in the PluginEditor which edits the build.bnd. In this case when there is a macro like ${.} or ${build} we want to see it in its raw unprocessed form (instead of the actual expanded value).

This change makes sure that the methods used by the UI (BndEditorModel.getPluginsProperties() and BndEditorModel.setProperties()) will not expand macros

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>