1st draft: fix source lookup for bundle classpath lib jars
this is a first rough sketch which solves the problem that the source lookup during a debug session did not work when going into a method (F5) of a class which was coming from a jar inside the bundle classpath which was included via -includeresource: ${repo;bsn;latest}; lib:=true,\

I found out that the reason was that the reason was, that jars in the Bundle Classpath are not considered for source lookup.

F3 worked fine though, but that does not help while debugging, because you want to step through the code.

this first commit is rough and needs more filtering and fine tuning, since I am adding too many jars. I need to figure out how to select only the few jars added via -includeresource

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>