refresh the EclipseWorkspaceRepository in RepositoriesViewRefresher
This is an improvement of commit 0e1fac345.
We now hook into the RepositoriesViewRefresher which already gets triggered after projects have been built to update the Repositories list.
The only thing which is missing here is a Refresh of the EclipseWorkspaceRepository, which is needed so that new "Project template projects" show up immediately in the "New Project" wizard

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>