RepoView: Clipboard: Bundles with self-imports
For debugging purposes it is useful to identify bundles containing self-imported packages.
I added this to the Repo-Browser on Right Click on a Repo / Copy to Clipboard / Bundles with Self-imports.
It outputs a simple list to the clipboard which can be pased to a texteditor for further analysis. In the future this might be refactored into more sophisticated solution but for the moment it is already useful.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>