bring back the no-projects links
now the links for
1. Create a new bnd workspace.
2. Bnd OSGi Workspace (Deprecated)

appear again in the Bnd Perspective for an empty workspace if you have no projects.
unfortunately I had to shorten the description because it is used instead of the name and was too long.
Drawback is, that  the File / New Entries used that as a tooltip, which is now much shorter.
But at least both views look ok

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>