move logic to MetaInfServiceParser
it leverages the recent addition of PR https://github.com/bndtools/bnd/pull/5912 that allowed to add annotations in comments of files in META-INF/services. But in this PR we basically pretent and add a aQute.bnd.annotation.spi.ServiceProvider annotation artificially. this causes bnd to generate Provide-Capability manifest headers for the services

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>