cnf change: prevent build disabling 'Build automatically'
This improves the DX in the following scenario:
When you edit /cnf/build.bnd and you Save the file before this fix a FULL build was always triggered, even if Project / Build Automatically was disabled. (and such a build can take very long and can become anoying in cases where you need tweak build.bnd)
Now a build is only triggered if Build Automatically  is enabled. This at least gives the developer a knob to disable the build-on-change behavior in cases where you work on the file and want to save it a few times along the way.

Note: The "Reload Workspace" Button in the Bndtools Explorer is not affected by this change and still triggeres a FULL build regardless of the Build Automatically setting (which is desired, because this Reload Button should act like a 'force build'

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>

Revert "cnf change: prevent build disabling 'Build automatically'"

This reverts commit dcd301e6457f188c8300283a3d7c0e0bcc072265.

cnf change: prevent build when 'Build automatically' is disabled

This improves the DX in the following scenario:
When you edit /cnf/build.bnd and you Save the file before this fix a FULL build was always triggered, even if Project / Build Automatically was disabled. (and such a build can take very long and can become anoying in cases where you need tweak build.bnd)
Now a build is only triggered if Build Automatically is enabled. This at least gives the developer a knob to disable the build-on-change behavior in cases where you work on the file and want to save it a few times along the way.

Note: The "Reload Workspace" Button in the Bndtools Explorer is also affected by this change and does not trigger a FULL build. This behavior is useful so that validation of cnf/build.bnd kicks in and you see e.g. Problem markers, without triggering a build

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>