Derive source file from builder when processing DS annotations
Currently if one uses the bnd-maven-plugin and has a component that
contains an error this is annotated at the pom.xml file as an error what
is confusing.

This now checks if an error is created if the source file to the class
can be found in the builder and then uses that as the file for reporting
the error. Due to bnd operating on the class files we still can't know
the exact line number (would require to parse it e.g. with JDT AST) but
having the actual file annotated already helps to identify the error
faster.

Signed-off-by: Christoph Läubrich <laeubi@laeubi-soft.de>