sub-bundles: show warning marks in sub .bnd file
instead of project bnd.bnd

This fixes the problem that previously, when you had sub-subbundles, then Warning / Error Markers where always shown at the parent bnd.bnd in the project, instead of in the sub-bundle's .bnd file (where the warning actually belongs to).

Now it is showing up correctly in the sub-bundle's .bnd file if it is comming from a sub-bundle, or if it is a normal bundle, it shows at the bnd.bnd as it already was the case.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>

move methods to Processor