Add a ProjectLauncher#getCommand for easier extension
Currently if one wants to extend a Project launcher it is required to
completely reimplement the launch() method even if one only wants to add
some argument to the commandline. Also things like returncode handling
and reporting of results need to be implemented again.

This now adds a new method ProjectLauncher#getCommand() that externalize
the process of creating a commandline, this also allows to access the
final commandline from the outside (e.g. to show it to the user).

Signed-off-by: Christoph Läubrich <laeubi@laeubi-soft.de>