hint that Eclipse restart is required
adjust description of the warning, so at least it is mentioned somewhere.
unfortunatelly I could not figure out how to reliably sync the Processor.pendatic field with the .bnd files on change. The problem is that Processor.pendatic can be set via .bnd files but also via Processor.setPedantic(boolean) e.g. from testcases, there is some magic caching going on in Processor.begin() and Processor.getProperties() which I was not able to figure out, what to call so that the Workspace Refresh Button behaves the same as an Eclipse restart.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>

Update pedantic.md