add PackageRef.isJDK(EE)
Experimental: this allows us to check if a given packageRef is part of the JDK for the given EE.
If true then we suppress the pedantic warning "Imports that lack version ranges"
TODO this current breaks test.BuilderTest.testRemovedImportWithRequireBundle() which explicityl expects the warning "Imports that lack version ranges: [javax.swing]". Need to see if I can build a replacement.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>