use Eclipse 2022-09 simrel p2 site
instead of handcrafted at https://bndtools.jfrog.io/bndtools/bnd-build/eclipse/4.25/index.xml.gz

This is a first step of easier switching to a newer Eclipse Versions in the future, since we start having more and more bugs or behavioral changes only happening in newer Eclipse versions.

- created a local P2 Index local /cnf/ext/eclipse.p2.xml.pobr index of Eclipse 4.25.0.20220908-1200 Release Repo (SimRel), which avoids a full download of the whole repo initially. only the .jar files which are really needed will be downloaded in a lazy fashion
- used the new .pobr index mechanism (see MagicBnd.java) to add a repository without configuring it as a plugin... just by naming the file with .pobr
- added description how to manually create this index to cnf/repositories.bnd
- in the future we could look for ways to simplify this process
- I also resolved the .bndrun files against the new repo... a few versions

- experimental: prepared github action Cache for caching /cnf/cache and ~/.bnd/urlcache/ to have similar caching behavior in CI that we have locally in Eclipse /cnf/cache
This is currently commented out in cibuild.yml, but we can enable it later if needed. But first we want to test the status quo

- re-resolved all main .bndrun files

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>