keep duplicate -runbundles
this fixes the problem that in a RunResultion which suggested the same bundle but in different versions only one bundle (the highest) was kept. Now both suggestions are kept.

The problem was that Parameters.toParameter() used parameters.put() instead of parameters.add(). .put() removes duplicate keys, .add() keeps duplicates by using the duplicate markers.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>