bndrun: refresh model before Resolve
this fixes the problem that the Resolve button did not "see" changes of included .bndrun files (e.g. -include: shared.bndrun). So when you changed something in an include and pressed the "Resolve" button , then the "Resolve" process would still see the old state when the BndEditor was initially opened and as if the change to -include: shared.bndrun would not have happened. Previous workaround was to close and open the BndEditor of the actual .bndrun file again

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>