supress warning for optional Imports w/o version
This suppresses the pedantic warning "Imports that lack version ranges" for the case of resolution:=optional imports. The idea behind this is, that if the developer has made the effort to customize the imports, then the developer has "handled" or "touched" that package. And one could argue the developer could have added a version too. But since the developer decided to only add resolution:=optional, then the version is probably not known or not important

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>

Revert "supress warning for optional Imports w/o version"

This reverts commit 1616420eb4f070bbbefcd6033631e2aea130b58a.

add more info to warning