fix Customize Imports forgets selection after save
this fixes the bug that after saving e.g. a change to the optional checkbox did not have an effect immediatelly and the UI element of the checkbox was not in sync with the actual values in the source tab and also in the table viewer

Now after the fix the state of the UI Elements and source are always in sync after saving.
Bonus: the current selection of values (multiple selections) is remembered.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>