create per-project empty.index
During debugging I found why it is still not working. I found the index is build parallel in multiple threads and if there is a single shared empty.index file the threads have a problem with this, close the empty.index and start an (potentially in memory) reindexing of the actual .jar (which is what we want to prevent). So giving each project its own empty-projectname.index we don't run into race conditions of a shared empty.index

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>