allow excluding repositories from Project -buildpath via tags
e.g. only consider "resolve" repos for Project. getRepositories()

This fixes a cornerish-case problem with compilation errors when you temporarily add a new Repository to  your workspace which is purely for lookup/research/browsing purposes and should neither be considered for resolution nor to the buildpath or Eclipse Classpath. It should just "be there" but do nothing.

It fixes a problem I had when I temporarily added an Eclipse Repo for the research purpose above and suddenly got compile errors because the new repo contained a 2.x version of slf4j.simple while the code expected to get an 1.7.x via -buildpath: slf4j.simple;version=latest

But maybe this is a practical feature, not only for this corner case.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>