support source attchmts for sub-bundles + ${repo} refs
This makes the Eclipse Source Attachments work also for sub-bundles. This is especially useful for  pure wrapper bundles which contains only sub-bundles to wrap non-osgi jars and all projects in the workspace only have a build-path reference to the wrapper bundle, but not to the wrapped actual jars (which have the -sources.jar in .m2 repo).

Also improved how the version is used, but also using Container.getVersion() in case c.getAttributes().get(Constants.VERSION_ATTRIBUTE) is null. This makes it work for cases where a ${repo} reference specifies an explicit version (not latest) e.g. -includeresource: ${repo;org.apache.lucene:lucene-analyzers-common;[4.0.0,4.0.1)};lib:=true

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>