Copilot _heads/_ext
Model: GPT-4.1
Prompt: The current .md files in the _heads/_ext folder of the bnd / bndtools OSGi project contains  code snippets as the content. This was a hack / workaround from the past in hope somebody would write this as prose. Please do so now. The frontmatter block might give clues.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>