bnd build now does a compile + build
because `bnd compile` alone just compiles .java files to class files but will fail for workspace bundles which require another bundle (because compile expects the built .jar  of that dependent bundle on the classpath, but the .jar is not there unless bnd build has created the jar file). So kind of a henn-egg problem.

- add project.generate()
- deprecate 'bnd compile' for removal in 8.0

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>