add new cli command: 'bnd dev app.bndrun'
This is combines 'bnd build --watch & bnd run myapp.bndrun' in a single command.
It provides a Live Coding experience on the command line as known from Eclipse with Bndtools. This allows to use another IDE e.g. vscode or a normal text editor. The command should be executed the workspace or the project containing the .bndrun

Example:
```
cd bndworkspace
bnd dev appbundle/myapp.bndrun
```

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>