bndtools: fix Eclipse sourcelookup for runfw
This allows to view the source code attachment for classes in the Runframework. For example I needed to debug org.eclipse.osgi.internal.serviceregistry.ServiceFactoryUse which did not find the sourcecode, although I have it in my local ~/.m2 maven repo and (downloaded via bndtools repo browser).

This change fixes it by using the same lookup mechanism as for other bundles

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>