log Already closed instead throw
now we log.warn the Already closed Exception of MavenBndRepo and just return false, as the method does in other cases too.

I don't really know why the exception was thrown in the first place. The thing is that .init() is constantly called by polling (see .startPoll()) or other Workspace refresh actions which have references to the repo.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>