show the bnd-cache in repo browser
this makes the previously hidden repo visible in the repo-browser.
It helps devs to know what's going on behind the scenes e.g. because this bnd-cache contains the ProjectLauncher which is automatically added to exportable .jars. But so far it has been unclear where this launcher is magically coming from. So this is an attempt to make this repo visible.

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>