use P2Exporter for org.bndtools.p2
first draft of using https://bnd.bndtools.org/plugins/p2export.html for bnd itself
the goal is to replace the veery old way of building the p2 artifact and use bnd itself.

Importantly this solves the issue to use SHA-256 for checksum in addition to md5 to solve the Eclipse warning like "The digest algorithms (md5) used to verify ...jar have severely compromised security. Please report this concern to the artifact provider."

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>