do not concat BUNDLE_ACTIVATOR
fixes a problem in Verifier.verifyActivator() ("The Bundle-Activator header only supports a single type. The following types were found") when the bnd instructions already define a Bundle-Activator but a processed jar also contributed a Bundle-Activator. Before this fix the Bundle-Activator header was concatenated to the existing Bundle-Activator header in bnd.bnd and thus contained multiple values (separated by comma). But for Bundle-Activator this does not make sense as it supports only a single value.

Added Constants.BUNDLE_HEADERS_SINGLE_VALUE for a list of other headers which should also not be concatenated

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>